added test for cells, split tests into respective directories
